requirements.txt
pandas
chromadb
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-openai
openai
llama-index-embeddings-huggingface
gradio
mammoth
requests
